refactor: (#26) FindAccountId UseCase human error
